Change `short` field from string to boolean
The Slack documentation establishes this should be boolean
https://api.slack.com/docs/message-attachments

While Slack does accept the string value, local testing shows
that it may well always interpet it as 'False'. As such, we
default to True to match the existing behaviour.

This also allows the plugin to notify Mattermost,
which is compatible with Slack webhooks.

Closes #5